fix: recommendations hooks
