+++ src/widgets/RecommendationsPanel/components/CourseCard.jsx
@@ -3,33 +3,13 @@ import PropTypes from 'prop-types';
 
 import { Card, Hyperlink, Truncate } from '@edx/paragon';
 
-import { configuration } from 'config';
-import { setCookie, getCookie } from 'utils/cookies';
-
 import { useIsCollapsed } from 'containers/CourseCard/hooks';
-import track from '../track';
+import useCourseCardData from './hooks';
 import './index.scss';
 
 export const CourseCard = ({ course, isPersonalizedRecommendation }) => {
   const isCollapsed = useIsCollapsed();
-
-  const handleCourseClick = (e) => {
-    e.preventDefault();
-    const cookieName = configuration.PERSONALIZED_RECOMMENDATION_COOKIE_NAME;
-    let recommendedCourses = getCookie(cookieName);
-    if (typeof recommendedCourses === 'undefined') {
-      recommendedCourses = { course_keys: [course.courseKey] };
-    } else if (!recommendedCourses.course_keys.includes(course.courseKey)) {
-      recommendedCourses.course_keys.push(course.courseKey);
-    }
-    recommendedCourses.is_personalized_recommendation = isPersonalizedRecommendation;
-    setCookie(cookieName, JSON.stringify(recommendedCourses), 365);
-    track.recommendedCourseClicked(
-      course.courseKey,
-      isPersonalizedRecommendation,
-      course?.marketingUrl,
-    );
-  };
+  const { handleCourseClick } = useCourseCardData(course, isPersonalizedRecommendation);
 
   return (
     <Hyperlink
